Actioning review issue by migrating change log code from ReferralService > AmendReferralService and adjust tests.
