+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/ReferralServiceUnitTest.kt
@@ -56,6 +56,7 @@ class ReferralServiceUnitTest {
   private val referralAccessFilter: ReferralAccessFilter = mock()
   private val userTypeChecker: UserTypeChecker = mock()
   private val communityAPIOffenderService: CommunityAPIOffenderService = mock()
+  private val amendReferralService: AmendReferralService = mock()
   private val hmppsAuthService: HMPPSAuthService = mock()
   private val telemetryService: TelemetryService = mock()
   private val referralDetailsRepository: ReferralDetailsRepository = mock()
@@ -82,6 +83,7 @@ class ReferralServiceUnitTest {
     serviceProviderAccessScopeMapper,
     referralAccessFilter,
     communityAPIOffenderService,
+    amendReferralService,
     hmppsAuthService,
     telemetryService,
     referralDetailsRepository,
